Fix leftover code in FileReader destructor from 1f46b1ed48dc35f96282d3f6ff420fcd39287732
Previously, the FileReader checked if the file was open and if and only
if, closed it, alongside deleting the underlying file stream. However,
if the file stream was existent, but the file not open, it did not
delete the file stream, resulting in a memory leak.

Fix this by deleting the check for if the file is open, so it only
checks if the file stream is existent, and letting another if statement
close the file inside it. After that, always delete the file stream.