Make FileReader extend std::ifstream
So mappings for every single member of std::ifstream for FileReader
don't have to be created, just make FileReader inherit std::ifstream.

Note that this automatically declares the copy constructor of FileReader
as deleted. In order to copy FileReader A to FileReader B, get the path
of FileReader A and construct a FileReader object B with the path as
it's argument