Fix TokenParser always going out-of-range
The TokenParser always went out-of-range. This is due to the fact, that
in it's pointer incrementing member functions, it was checked if the
resulting increment would make the pointer out-of-range. This makes
parsing impossible, since the member function to check if the
pointer is out of range will be useless. Remove the check.